requirements.txt
crewai>=0.34.5
langchain>=0.0.316
langchain-community>=0.0.12
fastapi>=0.110.0
uvicorn>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.5.2
openai>=1.12.0
requests>=2.31.0
pyyaml>=6.0.1
aiohttp>=3.9.3